fix: removed gui
