+++ Cargo.toml
@@ -33,11 +33,5 @@ memflow = { version = "=0.2.0-beta10", features = ["plugins"] }
 serde = { version = "1.0", features = ["derive"] }
 toml = "0.7"
 
-# gui
-glium = "0.29.0"
-imgui = "0.7.0"
-imgui-glium-renderer = "0.7.0"
-imgui-winit-support = "0.7.0"
-
 [target.'cfg(windows)'.build-dependencies]
 winres = "0.1"
