--- src/gui/mod.rs
@@ -1,6 +1,0 @@
-mod support;
-
-pub mod settings;
-pub use settings::{Config, Settings};
-
-pub mod alert;
