+++ src/memflow_wrapper.rs
@@ -1,7 +1,4 @@
-use crate::gui::{
-    alert,
-    settings::{Config, Settings},
-};
+use crate::config::settings::{Config, Settings};
 use std::collections::HashMap;
 use std::sync::{Arc, Mutex, MutexGuard};
 
@@ -18,11 +15,6 @@ pub unsafe fn lock_memflow<'a>() -> Result<MutexGuard<'a, Memflow>> {
                 MEMFLOW_INSTANCE = Some(Arc::new(Mutex::new(memflow)));
             }
             Err(err) => {
-                alert::show_error(
-                    "Unable to load memflow",
-                    "Memflow failed to initialize some of its components",
-                    err,
-                );
                 return Err(err.log_error("unable to initialize memflow"));
             }
         };
@@ -55,18 +47,9 @@ impl Memflow {
         simple_logging::log_to_file("memflow_reclass.log", LevelFilter::Info).ok();
 
         // load config file and set initial logging level
-        let mut settings = Settings::new();
+        let settings = Settings::new();
         log_level_from_str(settings.config().log_level.as_ref());
 
-        // show configuration dialog
-        settings.configure();
-        if let Err(err) = settings.persist() {
-            alert::show_error(
-                "Unable to save settings",
-                "The configuration file could not be written",
-                err,
-            );
-        }
         let config = settings.config();
 
         // update logging level after showing the configuration dialog
